+++ NEWS.md
@@ -1,5 +1,6 @@
 ## xxxx-xx-xx 6.1.0
 * [MODSOURCE-1014](https://folio-org.atlassian.net/browse/MODSOURCE-1014) Fix 035 being split on update when same $a in incoming and existing
+* [MODDICORE-509](https://folio-org.atlassian.net/browse/MODDICORE-509) Fix MARC-MARC matching when comparison part is used without a qualifier
 
 ## 2026-04-15 6.0.0
 * [MODSOURCE-917](https://folio-org.atlassian.net/browse/MODSOURCE-917) "Numerics only" option of existing record section does not work during MARC-BIB to MARC-BIB matching